fix: error en proceso de fecha por JSON, tasa-interes-pasiva

+++ sicf-mantenimientos-generales/src/main/java/ob/unibanca/sicf/mantenimientosgenerales/model/TasaInteresPasiva.java
@@ -25,6 +25,6 @@ public class TasaInteresPasiva {
     private Float tasaInteresMonedaBase;
     private Float tasaInteresMonedaAlterna;
 
-    private Float factorDiarioMonedaBase;
-    private Float factorDiarioMonedaAlterna;
+    private Double factorDiarioMonedaBase;
+    private Double factorDiarioMonedaAlterna;
 }
